Add Alarm use case basics
New approach:
I'll try to complete a use case completely (at least up to showing up in
the GUI) before going to the next one.
Starting here with the work on adding an alarm and seeing it added;
removed some code that wasn't necessary for this, though it will
probably be added again later.